// frontend/src/app/apiClient/model/caseEscortReceipt.ts
/**
 * LoopBack Application
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



import { BigDecimal } from 'bigdecimal';




export interface CaseEscortReceipt { 

    
    receiptID: string;
    
    caseID: string;
    
    escortID: string;
    
    name: string;

    alternateName?: string;
    
    storageHash?: string;
    
    createDate: Date;
    
    currencyType?: string;
    
    amount?: BigDecimal;
    
    receiptDate?: Date;

    usdAmount?: BigDecimal;
}